% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.episode.list.to.record}
\alias{add.episode.list.to.record}
\title{Adding a list of ccEpisode to ccRecord}
\usage{
add.episode.list.to.record(rec, lst)
}
\arguments{
\item{rec}{ccRecord2}

\item{lst}{a list of ccEpisode2 objects}
}
\value{
ccRecord2
}
\description{
Adding a list of one or multiple ccEpisode2 objects to
ccRecord2 object, the information table (infotb) will be updated automatically.
It is a more efficient way to add multiple ccEpisode2 objects. See
add.episode.to.record() for just adding one ccEpisode.
}

