% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.episode.to.record}
\alias{add.episode.to.record}
\title{adding episode data as a list to ccRecord}
\usage{
add.episode.to.record(rec, episode)
}
\arguments{
\item{rec}{ccRecord2}
}
\description{
adding episode data as a list to ccRecord
}

