% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutivetime.r
\name{add}
\alias{add}
\title{Helper functions to add the numbers on an array in a
cummulative fashion using Reduce
extracted from: http://adv-r.had.co.nz/Functionals.html}
\usage{
add(x, y, na.rm = FALSE)
}
\description{
Helper functions to add the numbers on an array in a
cummulative fashion using Reduce
extracted from: http://adv-r.had.co.nz/Functionals.html
}

