% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{addEpisode,ccPatient,ccEpisode-method}
\alias{addEpisode,ccPatient,ccEpisode-method}
\title{addEpisode}
\usage{
\S4method{addEpisode}{ccPatient,ccEpisode}(obj, episode)
}
\description{
addEpisode
}

