% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{addEpisode,ccPatient,cEpisode-method}
\alias{addEpisode,ccPatient,cEpisode-method}
\title{addEpisode}
\usage{
\S4method{addEpisode}{ccPatient,cEpisode}(obj, episode)
}
\description{
addEpisode
}

