% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccPatient.r
\name{addEpisodeListToPatient}
\alias{addEpisodeListToPatient}
\title{add a list of episode to a ccPatient object}
\usage{
addEpisodeListToPatient(patient, eplist)
}
\arguments{
\item{patient}{ccPatient}

\item{eplist}{list contains ccPatient objects}
}
\description{
add a list of episode to a ccPatient object
}

