% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccPatient.r
\name{addEpisodeToPatient}
\alias{addEpisodeToPatient}
\title{add ccEpisode to ccPatient object}
\usage{
p <- ccPatient()
e <- ccEpisode()
p_new <- p + e
}
\arguments{
\item{patient}{ccPatient}

\item{episode}{ccEpisode}
}
\value{
ccPatient
}
\description{
add ccEpisode to ccPatient object
}

