% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdata.r
\name{addEpisodeToRecord}
\alias{addEpisodeToRecord}
\title{add ccEpisode object to ccRecord object}
\usage{
r_new <- r + e
}
\arguments{
\item{recd}{ccRecord}

\item{episode}{ccEpisode}
}
\value{
ccRecord
}
\description{
add ccEpisode object to ccRecord object
}

