% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{addItemData,cEpisode,data.frame-method}
\alias{addItemData,cEpisode,data.frame-method}
\title{as 1D data
as 2D data}
\usage{
\S4method{addItemData}{cEpisode,data.frame}(obj, data)
}
\description{
as 1D data
as 2D data
}

