% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccEpisode.R
\name{addItemData}
\alias{addItemData}
\title{add data in a data frame to an episode.}
\usage{
addItemData(obj, data)
}
\arguments{
\item{obj}{episode}

\item{data}{data frame with either 2 columns (id, val) or 3 columns}
}
\description{
add data in a data frame to an episode.
}

