% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findMaxMinTime.R
\name{addPseudoTime}
\alias{addPseudoTime}
\title{This function is to assign a pseudo time, which is the earliest and latest
appeared values of time-series items to admission and discharge time
respectively to a record. Please only use this when dealing with the anonymised
data, as it modifies the dataset itself.}
\usage{
addPseudoTime(record)
}
\description{
This function is to assign a pseudo time, which is the earliest and latest
appeared values of time-series items to admission and discharge time
respectively to a record. Please only use this when dealing with the anonymised
data, as it modifies the dataset itself.
}

