% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdata.r
\name{addRecord}
\alias{addRecord}
\title{append two ccRecord objects}
\usage{
r_new <- r1 + r2
}
\arguments{
\item{rec1}{ccRecord}

\item{rec2}{ccRecord}
}
\value{
ccRecord
}
\description{
append two ccRecord objects
}

