% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{aggPatients,ccRecord-method}
\alias{aggPatients,ccRecord-method}
\title{Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)}
\usage{
\S4method{aggPatients}{ccRecord}(record)
}
\arguments{
\item{record}{ccRecord}
}
\description{
Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)
}

