% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{allData}
\alias{allData}
\title{Return all the the recorded data.
This is a connection to a database / query,
not a real value. If you want to use it you will
need to run collect() on it.}
\usage{
allData(connection)
}
\description{
Return all the the recorded data.
This is a connection to a database / query,
not a real value. If you want to use it you will
need to run collect() on it.
}

