% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{allPatientsInfo}
\alias{allPatientsInfo}
\title{Retrieve one item value at a time from all patients.}
\usage{
all.nhs.number <- allPatientsInfo(ccd, "nhs_number")
}
\arguments{
\item{record}{is a ccRecord object.}

\item{item}{is a character which can be "nhs_number", "pas_number" and
"site_id"}
}
\value{
a vector a all patients selected item.
}
\description{
Retrieve one item value at a time from all patients.
}

