% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stdid.r
\docType{methods}
\name{as.number,StdId-method}
\alias{as.number,StdId-method}
\title{convert standard ids to numbers (character) which can be used for indexing.}
\usage{
\S4method{as.number}{StdId}(obj)
}
\description{
convert standard ids to numbers (character) which can be used for indexing.
}

