% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{ccd_demographic_spell}
\alias{ccd_demographic_spell}
\title{Create demographic table with spell IDs}
\usage{
ccd_demographic_spell(rec, duration = 2)
}
\arguments{
\item{rec}{ccRecord}

\item{duration}{the maximum hours of transition period}
}
\value{
data.table demographic table with spell ID in column spell
}
\description{
same output like ccd_demographic_table but in 
addition with a spell ID.
}

