% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{ccd_demographic_table}
\alias{ccd_demographic_table}
\title{Create the demographic tables, which includes all non-time-varying variables.}
\usage{
ccd_demographic_table(record, dtype = TRUE)
}
\arguments{
\item{record}{ccRecord-class}

\item{dtype}{logical column will be type aware, else all in character.}
}
\description{
The data type of each column is in its corresponding data 
type.
}

