% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{ccd_select_table}
\alias{ccd_select_table}
\title{Create the table for ccTable from ccRecord}
\usage{
ccd_select_table(record, items_opt = NULL, items_obg = NULL, freq,
  return_list = FALSE)
}
\arguments{
\item{record}{ccRecord}

\item{items_opt}{character vectors. Items (HIC code) selected in item_opt are optional items, which will be automatically 
filled when item is missing.}

\item{items_obg}{obligatory items that is obligatory; Any episode that does not contain
item in this vector will be removed.}

\item{freq}{numeric cadence in hour.}

\item{return_list}{logical if TRUE return as a list.}
}
\value{
data.table
}
\description{
Create the table for ccTable from ccRecord
}

