% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csv2ccdata.r
\docType{class}
\name{ccdata-class}
\alias{ccData}
\alias{ccdata-class}
\title{\section{Slots}{
   \describe{
     \item{\code{a}:}{Object of class \code{"numeric"}.}
     \item{\code{b}:}{Object of class \code{"character"}.}
   }
 }}
\description{
\section{Slots}{
   \describe{
     \item{\code{a}:}{Object of class \code{"numeric"}.}
     \item{\code{b}:}{Object of class \code{"character"}.}
   }
 }
}

