% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccDataTable2.R
\docType{class}
\name{ccDataTable2-class}
\alias{ccDataTable2}
\alias{ccDataTable2-class}
\title{this is ref example}
\description{
this is ref example
}
\section{Methods}{

\describe{
\item{\code{create.table(freq)}}{Create a table contains the selected items in the conf with a given
frequency (in hour)}

\item{\code{filter.null(items = c("episode_id", "site"))}}{remove the entire episode when the episode_id or site is NULL}

\item{\code{reload.conf(file)}}{reload yaml configuration.}
}}

