% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{class}
\name{ccEpisode-class}
\alias{ccEpisode}
\alias{ccEpisode-class}
\title{The S3 class which holds data of a single episode.}
\description{
The S3 class which holds data of a single episode.
}
\section{Fields}{

\describe{
\item{\code{site_id}}{character string. Site ID, if presented, otherwise "NA".}

\item{\code{episode_id}}{character string. Episode ID, if presented, otherwise "NA".}

\item{\code{nhs_number}}{character string. NHS number, if presented, otherwise "NA".}

\item{\code{pas_number}}{character string. PAS number, if presented, otherwise "NA".}

\item{\code{parse_file}}{character string. The source XML file. If the source is not a file then "NA".}

\item{\code{t_admission}}{POSIXct. Time of Admission to the ICU, if presented, otherwise NA.}

\item{\code{t_discharge}}{POSIXct. Time of discharge of the ICU, if presented, otherwise NA.}

\item{\code{parse_time}}{POSIXct. Parse time.}

\item{\code{data}}{A list which holds all the data of this episode which is indexed by NIHIC code.}
}}

