% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccPatient.r
\docType{class}
\name{ccPatient-class}
\alias{ccPatient}
\alias{ccPatient-class}
\title{Patient class data}
\description{
Patient class data
}
\examples{
p <- ccPatient() #initialise an empty patient.
}

