% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/record2table.r
\name{ccRecord2Table}
\alias{ccRecord2Table}
\title{Convert ccRecord data to item tables}
\usage{
ccRecord2Table(record, items = NULL, pseudo = FALSE, file = NULL)
}
\arguments{
\item{record}{ccRecord}

\item{items}{if specified, items should be the NHIC code of the selected
items, if not the output will contain be all the possible items, even
those not appearing in the ccRecord}

\item{pseudo}{(optional) logical value, FALSE gives identifier the PAS number, TRUE gives
the patient count in the record.}

\item{file}{the name of csv files. If speicified 1d/2d tables will be
written into two sperate csv files.}
}
\description{
Convert ccRecord data to item tables
}

