% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{ccTable_apply_filters}
\alias{ccTable_apply_filters}
\title{Apply all the setup filters.}
\arguments{
\item{warnings}{logical value to indicate more or less messages with an 
default value TRUE.}
}
\description{
Once filters are applied, the processed data will be stored in tclean. Note, 
running filtering function before apply_filters is necessary. This function 
will have no effect on tclean if no filter is ran prior.
Filters will decide to preserve or remove particular entries or episodes.
}
\examples{
\dontrun{
tb <- create_cctable(ccd, conf, 1)
tb$range_filter() 
tb$apply_filter() # apply only the range filter ragardless of the conf. 
}
}

