% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{ccTable_create_cctable}
\alias{ccTable_create_cctable}
\title{Create a ccTable object}
\description{
This is a member function of ccTable-class. Using create_cctable is a safer and 
easier way to create the ccTable. See create_cctable.
}

