% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.missingness.R
\name{ccTable_filter_nodata}
\alias{ccTable_filter_nodata}
\title{No data filter}
\description{
Remove the episode when a particular field is not presented.
It need to be set up in the YAML configuration file.
}

