% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{ccTable_reload_conf}
\alias{ccTable_reload_conf}
\title{Reload the YAML configuration file}
\description{
Note, this function will also reset all the operations and 
remove the tclean.
}
\examples{
\dontrun{
tb$reload_conf()
}
}

