% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sanity.check.r
\name{check.sanity}
\alias{check.sanity}
\title{0 - unaccepted
1 - unusual
2 - accepted
3 - normal}
\usage{
check.sanity(record, ref, verbose = FALSE)
}
\description{
0 - unaccepted
1 - unusual
2 - accepted
3 - normal
}

