% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggPatients.r
\name{combineEpisode}
\alias{combineEpisode}
\title{Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)}
\usage{
combineEpisode(record)
}
\description{
Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)
}

