% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score.r
\name{computeItemScore}
\alias{computeItemScore}
\title{compute one item of the socre of a single patient.}
\usage{
computeItemScore(data.set, patient.id, item.code, range, score.ref, fun)
}
\arguments{
\item{data.set}{}

\item{patient.id}{}
}
\value{
score in integer.
}
\description{
compute one item of the socre of a single patient.
}

