% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/computeMortality.r
\name{computeMortalityRate}
\alias{computeMortalityRate}
\title{compute mortality rate in each score from a data.frame which contains scores
and mortality.}
\usage{
computeMortalityRate(score.mort)
}
\arguments{
\item{score.mort}{}
}
\value{
score, mortality rate.
}
\description{
compute mortality rate in each score from a data.frame which contains scores
and mortality.
}

