% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score.r
\name{computePatientScore}
\alias{computePatientScore}
\title{Compute the score for single patient}
\usage{
computePatientScore(data.set, patient.id, items.code, ranges, score.ref,
  fun.list)
}
\arguments{
\item{items.code}{an array of items code to be concerned.}

\item{range}{list}
}
\value{
score
}
\description{
Compute the score for single patient
}

