% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score.r
\name{computeScore}
\alias{computeScore}
\title{compute scores of a given score system to multiple/all patients of a data
set.}
\usage{
computeScore(data.set, items.code, ranges, score.ref, fun.list,
  check.items.code, patient.id = NULL)
}
\arguments{
\item{data.set}{a list which contains data2d and data1d}

\item{items.code}{...}
}
\description{
compute scores of a given score system to multiple/all patients of a data
set.
}

