% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{connect}
\alias{connect}
\title{Make a connection to the Database that can be used for queries}
\usage{
connect(host = "localhost", username = NULL, password = NULL,
  database = NULL)
}
\description{
Make a connection to the Database that can be used for queries
}

