% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutivetime.r
\name{consecutivetime}
\alias{consecutivetime}
\title{Produces an index label that doesn't change for consecutive times}
\usage{
indices <- consecutivetime(c(T, F, T, T))
}
\arguments{
\item{array}{is a boolean array}
}
\value{
a vector with indices for when there is changes
}
\description{
Produces an index label that doesn't change for consecutive times
}

