% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{create.cctable}
\alias{create.cctable}
\title{Create the ccTable object from ccRecord}
\usage{
create.cctable(rec, freq, conf = NULL)
}
\arguments{
\item{rec}{ccRecord}

\item{freq}{the data cadence in hour.}

\item{conf}{either the path of YAML configuration file or the configuration
structure in list.}
}
\value{
ccTable object
}
\description{
Create the ccTable object from ccRecord
}

