% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{create.fat.table}
\alias{create.fat.table}
\title{Fat table}
\usage{
create.fat.table(db, frequency = 1)
}
\description{
We should be able to give the users choices of how to align the data, whether by 
using ICU admission time as the baseline, or hospital admission, or even with the earliest 
and the latest data point. I have updated reallocateTime__ which allows such 
operations.
}

