% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create2dclean.r
\name{create2dclean}
\alias{create2dclean}
\title{create2dclean}
\usage{
create2dclean(record, config, freq = 1, nchunks = 1)
}
\arguments{
\item{record}{ccRecord}

\item{config}{the full path of the YAML configuration file}

\item{freq}{table cadence}

\item{nchunks}{integer number}
}
\value{
A cleaned 2d wide table
}
\description{
create2dclean
}

