% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{create_cctable}
\alias{create_cctable}
\title{Create a ccTable object}
\usage{
create_cctable(rec, conf = NULL, freq = 1)
}
\arguments{
\item{rec}{ccRecord}

\item{conf}{either the path of YAML configuration file or the configuration}

\item{freq}{a unique sampling frequency (in hours) for all variables. e.g. if freq is set to 
1, each row in ccTable will represent a record of one hour.}
}
\value{
ccTable
}
\description{
Re-arrange the ccRecord object to table format where each column stands 
for a variable and each row a record data point. The number of rows will 
depend on the sampling frequency set in this function. If the original data
has a higher recording frequency than the set frequency (freq), the closest 
data point will be taken. It is suggested the `freq` should not be set 
lower than the maximum recording frequency in the original dataset.
}

