% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csv2ccdata.r
\name{csv2ccd}
\alias{csv2ccd}
\title{read csv file and convert to ccdata}
\usage{
csv2ccd(file, ...)
}
\arguments{
\item{file}{input file location.}
}
\value{
ccdata type
}
\description{
read csv file and convert to ccdata
}

