% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{data.checklist}
\alias{data.checklist}
\title{This a reference table of NHIC data items.}
\description{
This a reference table of NHIC data items.
}
\author{
Sinan Shi \email{s.shi@ucl.ac.uk}
}
\keyword{data}

