% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deltaTime.r
\name{deltaTime}
\alias{deltaTime}
\title{convert calendar time data in a record to delta time comparing to the ICU
admission time.}
\usage{
deltaTime(record, anonymised = FALSE)
}
\description{
convert calendar time data in a record to delta time comparing to the ICU
admission time.
}

