% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{demg.distribution}
\alias{demg.distribution}
\title{demg.distribution
Create a plot of the distribution of numerical demographic data.}
\usage{
demg.distribution(demg, names)
}
\arguments{
\item{demg}{ccRecord or demographic table created by sql.demographic.table()}

\item{names}{character vector of short names of numerical demographic data.}
}
\description{
demg.distribution
Create a plot of the distribution of numerical demographic data.
}
\examples{
\dontrun{tdemg.distribution(ccd, "HCM")}
}

