% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{demographic.data.completeness}
\alias{demographic.data.completeness}
\title{create a demographic completeness table}
\usage{
demographic.data.completeness(demg, names = NULL, return.data = FALSE)
}
\description{
create a demographic completeness table
}

