% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.spell.R
\name{demographic.patient.spell}
\alias{demographic.patient.spell}
\title{Assign unique spell ID to the demographic table}
\usage{
demographic.patient.spell(rec, duration = 2)
}
\arguments{
\item{rec}{ccRecord}

\item{duration}{the maximum hours of transition period}
}
\value{
data.table demographic table with spell ID in column spell
}
\description{
Assign unique spell ID to the demographic table
}

