% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{demographic.table}
\alias{demographic.table}
\title{Create the demographic tables. The data type of each column is in its
corresponding data type.}
\usage{
demographic.table(record, dtype = TRUE)
}
\arguments{
\item{record}{ccRecord-class}

\item{dtype}{logical column will be type aware, else all in character.}
}
\description{
Create the demographic tables. The data type of each column is in its
corresponding data type.
}

