% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{demographics.table}
\alias{demographics.table}
\title{This gets you the demographics as a table out of your current data.}
\usage{
demographics.table(table)
}
\description{
This gets you the demographics as a table out of your current data.
}

