% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{episode.graph}
\alias{episode.graph}
\title{Individual episode graph}
\usage{
episode.graph(ccd, eid = 601, items = NULL)
}
\arguments{
\item{ccd}{ccRecord}

\item{eid}{character the episode index in the ccRecord}

\item{items}{character NIHC code of longitudinal data.}
}
\description{
Create an individual episode graph for its diagnosis, drugs and physiological
variables. Diagnosis and drugs are always included, while the user can
select other longitudinal data.
}
