% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{export.lontb}
\alias{export.lontb}
\title{Export all the longitudinal fields as tables}
\usage{
export.lontb(ccd)
}
\arguments{
\item{ccd}{ccRecord object}
}
\description{
All the result tables will be wrapped by a list, indexing with short names of
the fields.
}

