% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{exportData}
\alias{exportData}
\title{Return all the the recorded data but only fields users are allowed to see.
This is a connection to a database / query,
not a real value. If you want to use it you will
need to run collect() on it.}
\usage{
exportData(connection)
}
\description{
Return all the the recorded data but only fields users are allowed to see.
This is a connection to a database / query,
not a real value. If you want to use it you will
need to run collect() on it.
}

