% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2ccdata.R
\name{extract_file_origin}
\alias{extract_file_origin}
\title{Extract the original file name from a path and file removing
all the suffixes.}
\usage{
extract_file_origin(pathfile, removestr = ".xml")
}
\arguments{
\item{pathfile}{a particular file name which may have a suffix}

\item{removestr}{last bit from the original filename}
}
\value{
string
}
\description{
Extract the original file name from a path and file removing
all the suffixes.
}

