% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_info}
\alias{extract_info}
\title{Extract information from data.checklist}
\usage{
extract_info()
}
\value{
list of time [data.frame(id, idt)], meta [data.frame(id, idmeta)], 
        nontime [numeric], MAX_NUM_NHIC
}
\description{
Extract information from data.checklist
}
