% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extractIndexTable}
\alias{extractIndexTable}
\title{get indexing tables for time label, time-wise value, meta data label, and
meta data.}
\usage{
extractIndexTable()
}
\value{
list of vectors contains time.index, datat.index, meta.index,
         datam.index
}
\description{
get indexing tables for time label, time-wise value, meta data label, and
meta data.
}

