% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{file.summary}
\alias{file.summary}
\title{Produce a file summary table}
\usage{
file.summary(ccd)
}
\arguments{
\item{ccd}{ccRecord-class}
}
\value{
data.table
}
\description{
Produce a file summary table
}

