% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{for_each_episode}
\alias{for_each_episode}
\title{loop over all episodes of a ccRecord object}
\usage{
for_each_episode(record, fun)
}
\arguments{
\item{record}{ccRecord}

\item{fun}{function}
}
\description{
loop over all episodes of a ccRecord object
}

