% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csv2ccdata.r
\name{get.ccdata}
\alias{get.ccdata}
\title{read csv or xml file and convert to ccdata}
\usage{
get.ccdata(file, file.type = "", ...)
}
\arguments{
\item{file}{input file location.}

\item{file.type}{specify file type [optional].}
}
\value{
ccdata type
}
\description{
read csv or xml file and convert to ccdata
}

