% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCCData.r
\name{getCCData}
\alias{getCCData}
\title{read csv or xml file and convert to ccdata}
\usage{
getCCData(file, file.type = "", ...)
}
\arguments{
\item{file}{input file location.}

\item{file.type}{specify file type [optional].}
}
\value{
ccdata type
}
\description{
read csv or xml file and convert to ccdata
}

