% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{getEpisodePeriod}
\alias{getEpisodePeriod}
\title{Get the length of stay based on the first and the last data point.}
\usage{
getEpisodePeriod(e, unit = "hours")
}
\arguments{
\item{e}{ccEpisode object, represending a single episode.}

\item{unit}{character; units of time in which to return the output:
"secs", "mins", "hours", "days" or "weeks"}
}
\value{
a numeric of length 1, giving the length of stay
}
\description{
Get the length of stay based on the first and the last data point.
}
