% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getItemInfo}
\alias{getItemInfo}
\title{Retrieve information of the query code/item names from data.checklist}
\usage{
getItemInfo(item.code)
}
\arguments{
\item{item.code}{it can be either item name or NHIC_code, dt_code, or
meta_code}
}
\value{
a vector contains NHIC_code, dt_code, meta_code and row_in_checklist
}
\description{
Retrieve information of the query code/item names from data.checklist
}
\examples{
getItemInfo("Time of death on your unit")
getItemInfo("NIHR_HIC_ICU_0001")
}

