% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getItemsInfo}
\alias{getItemsInfo}
\title{get information of a group of code of items and return an array.}
\usage{
getItemsInfo(items.code, var)
}
\description{
get information of a group of code of items and return an array.
}

