% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPatientItems.r
\name{getPatient1dItem}
\alias{getPatient1dItem}
\title{retrieve one column of data from one or multiple patients.}
\usage{
getPatient1dItem(data1d, item.code, patient.id)
}
\arguments{
\item{patient.id}{}

\item{data2d}{}

\item{item}{}
}
\description{
retrieve one column of data from one or multiple patients.
}

