% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPatientItems.r
\name{getPatient2dItem}
\alias{getPatient2dItem}
\title{retrieve one column of 2d data from one or multiple patients}
\usage{
getPatient2dItem(data2d, item.code, patient.id)
}
\arguments{
\item{data2d}{}

\item{patient.id}{}

\item{item}{}
}
\description{
retrieve one column of 2d data from one or multiple patients
}

