% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{getXmlPatient}
\alias{getXmlPatient}
\title{get the patient data from xml}
\usage{
getXmlPatient(xml.root, id)
}
\arguments{
\item{xml.root}{root of xml data returned by xmlLoad()}
}
\value{
patient xml data
}
\description{
get the patient data from xml
}

