% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{icnarc2diagnosis}
\alias{icnarc2diagnosis}
\title{Convert the ICNARC code to human readable diagnosis}
\usage{
icnarc2diagnosis(icnarc)
}
\arguments{
\item{icnarc}{the ICNARC code, e.g. 1.1.1.1.1}
}
\value{
character ICNARC diagnosis
}
\description{
Convert the ICNARC code to human readable diagnosis
}

