% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icnarc_table}
\alias{icnarc_table}
\title{ICNARC diagnosis reference table}
\description{
ICNARC diagnosis reference table
}
\references{
\url{https://www.icnarc.org/Our-Audit/Audits/Cmp/Resources/Icm-Icnarc-Coding-Method}
}
\keyword{data}

