% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ITEM_REF.r
\name{importItemRef}
\alias{importItemRef}
\title{convert a yaml file to ITEMREF Rdata which will be loaded when the library
is called.}
\usage{
importItemRef(yml)
}
\description{
convert a yaml file to ITEMREF Rdata which will be loaded when the library
is called.
}

