% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccDataTable2.R
\name{inrange}
\alias{inrange}
\title{Check if the values of a vector v is in the ranges.}
\usage{
inrange(v, range)
}
\arguments{
\item{v}{vector numeric}

\item{range}{characters numeric ranges in a form such as (low, up). Multiple
ranges should seperated by semi-column.}
}
\description{
Check if the values of a vector v is in the ranges.
}

