% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{is.laboratory}
\alias{is.laboratory}
\title{Check if the item NHIC code or short name belongs to the Laboratory 
category.}
\usage{
is.laboratory(item_name)
}
\arguments{
\item{item_name}{character the NHIC code or the short name}
}
\value{
logical
}
\description{
Check if the item NHIC code or short name belongs to the Laboratory 
category.
}

