% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{is.longitudinal}
\alias{is.longitudinal}
\title{Check if the items NHIC code or short names belong is longitudinal.}
\usage{
is.longitudinal(item_name)
}
\description{
Check if the items NHIC code or short names belong is longitudinal.
}

