% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{left.join.var.table}
\alias{left.join.var.table}
\title{Join the existing vartb and fattb and create update the new fattb}
\usage{
left.join.var.table(dbname)
}
\arguments{
\item{dbname}{character database address}
}
\description{
Update the fattb and the vartb will be removed here.
}

