% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{lenstay}
\alias{lenstay}
\title{Calculate the length of stay in the ICU.}
\usage{
lenstay(demg, units = "hours")
}
\arguments{
\item{demg}{data.table the demographic table which should at least contain
column DAICU and DDICU}

\item{units}{character The unit of lenstay column, by default the output will be in hours}
}
\value{
data.table It is the original data.table with lenstay column (in 
difftime) appended.
}
\description{
Calculate the length of stay in the ICU and append it to the original demographic
table.
}

