% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{list_interpolation}
\alias{list_interpolation}
\title{interpolate list format}
\usage{
list_interpolation(l, item_id, lead = 1, lag = 1, FUN = mean, ...)
}
\arguments{
\item{lead}{}

\item{lag}{}

\item{FUN}{function}
}
\description{
interpolate list format
}

