% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{long2stname}
\alias{long2stname}
\title{Convert long names to short names.}
\usage{
long2stname(l)
}
\arguments{
\item{stname}{character long name such as "heart rate"}
}
\value{
short name character such as "h_rate"
}
\description{
Convert long names to short names.
}

