% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lookup.items}
\alias{lookup.items}
\title{Lookup items information by keywords}
\usage{
lookup.items(keyword, style = "grid")
}
\arguments{
\item{keyword}{character e.g. "h_rate", "heart", "108".}

\item{style}{character, the style of the table output which can be "simple",
"rmarkdown", and "grid"}
}
\value{
character the short names of the selected items.
}
\description{
This function tries to match keywords in short names, long names and NHIC code. 
The matched items will be displayed.
}

