% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{metadata}
\alias{metadata}
\title{The metadata table. This is all of the different 
codes that are available with their meanings, and
what the different columns mean for them.}
\usage{
metadata(connection)
}
\description{
This functions returns a loaded into memory R table,
**not** a connection to database table.
}
