% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccDataTable.R
\name{new.ccDataTable}
\alias{new.ccDataTable}
\title{This function reads first a yaml configuration file which cointains the
chosen items and their data quality information. It loads the selected items
and create a 2d missing rate table if the option check_miss2d is TRUE.}
\usage{
new.ccDataTable(record, yaml, base_frequency = 1, check_miss2d = TRUE)
}
\arguments{
\item{record}{is a ccRecord object of which the time of 2d data should be
numerical delta time.}

\item{yaml}{is a character type string specifying the location of the yaml
configuration file.}

\item{base_frequency}{integer frequency of the final table (in hour).}
}
\description{
This function reads first a yaml configuration file which cointains the
chosen items and their data quality information. It loads the selected items
and create a 2d missing rate table if the option check_miss2d is TRUE.
}

