% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{new.episode}
\alias{new.episode}
\title{Create a new episode}
\usage{
new.episode(lt)
}
\arguments{
\item{lt}{is a list}
}
\value{
ccEpisode object
}
\description{
create a new ccEpisode2 object by given the episode data as a
list. The list should be organised in data items. See examples.
}
\examples{
eps <- list()
eps[["NIHR_HIC_ICU_0018"]] <- data.frame(time=seq(10), rep(70, 10))
new.episode(eps)
}

