% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{parse.new.xml}
\alias{parse.new.xml}
\title{Update the RData database}
\usage{
parse.new.xml(xml.path, mc.cores = 4, quiet = FALSE)
}
\arguments{
\item{xml.path}{the path of the folder of which contains the XML files.}

\item{mc.cores}{number of processors to be applied for parallelisation.}

\item{quiet}{logical switch on/off of the progress bar.}
}
\value{
ccRecord object
combine new data to a ccRecord
}
\description{
Inject episode data from the newly added XML files to the RData database.
}

