% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggPatients.r
\name{patientIndex}
\alias{patientIndex}
\title{return selected index from nhs_number or pas_number}
\usage{
patientIndex(record, id)
}
\arguments{
\item{record}{ccRecord}

\item{id}{character, either the nhs_number or pas_number}
}
\value{
index vector
}
\description{
return selected index from nhs_number or pas_number
}

