% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{patientToDataArray}
\alias{patientToDataArray}
\title{convert xml data for a given patient and convert them into a list}
\usage{
patientToDataArray(patient, category.index.table)
}
\arguments{
\item{patient}{individual xml patient data}

\item{time.list}{}

\item{meta.list}{}
}
\value{
list
}
\description{
convert xml data for a given patient and convert them into a list
}

