% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{physio.distribution}
\alias{physio.distribution}
\title{plot the physiological data distribution.}
\usage{
physio.distribution(cctb, names)
}
\description{
plot the physiological data distribution.
}

