% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{plot,ccEpisode,character-method}
\alias{plot,ccEpisode,character-method}
\title{Episode chart}
\usage{
\S4method{plot}{ccEpisode,character}(r, v)
}
\arguments{
\item{r}{ccEpisode-class}

\item{v}{character}
}
\description{
Episode chart
}

