% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{plot,ccEpisode,missing-method}
\alias{plot,ccEpisode,missing-method}
\title{Episode chart default fields}
\usage{
\S4method{plot}{ccEpisode,missing}(r)
}
\arguments{
\item{r}{ccEpisode-class}
}
\description{
Episode chart default fields
}

