% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{plot_episode}
\alias{plot_episode}
\title{Individual episode chart}
\usage{
plot_episode(r, v)
}
\arguments{
\item{r}{ccEpisode-class}

\item{v}{short name of longitudinal data. While v is not given, the chart 
will only display h_rate, spo2, bilirubin, platelets, pao2_fio2, gcs_total.}
}
\value{
a table of selected vars of an episode
}
\description{
Create an individual episode chart for its diagnosis, drugs and physiological
variables. Diagnosis and drugs are always included, while the user can
select other longitudinal data.
}
\examples{
\dontrun{
plot_episode(ccd@episodes[[1]]) # plot first episode with default variables. 
plot_episode(ccd@episodes[[1]], "h_rate") # plot first episode heart rate
}
}

