% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccEpisode.r
\docType{methods}
\name{+,ccEpisode,data.frame-method}
\alias{+,ccEpisode,data.frame-method}
\title{add single or multiple record to episode.}
\usage{
\S4method{+}{ccEpisode,data.frame}(e1, e2)

\S4method{+}{ccEpisode,data.frame}(e1, e2)
}
\description{
add single or multiple record to episode.

add single or multiple record to episode.
}

