% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccPatient.r
\docType{methods}
\name{+,ccPatient,ccEpisode-method}
\alias{+,ccPatient,ccEpisode-method}
\title{overload the addEpisode to patient.}
\usage{
\S4method{+}{ccPatient,ccEpisode}(e1, e2)
}
\description{
overload the addEpisode to patient.
}

