% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{+,ccPatient,cEpisode-method}
\alias{+,ccPatient,cEpisode-method}
\title{overload the addEpisode to patient.}
\usage{
\S4method{+}{ccPatient,cEpisode}(e1, e2)
}
\description{
overload the addEpisode to patient.
}

