% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccdata.r
\docType{methods}
\name{+,ccRecord,ccEpisode-method}
\alias{+,ccRecord,ccEpisode-method}
\title{overload the addEpisode to patient.}
\usage{
\S4method{+}{ccRecord,ccEpisode}(e1, e2)
}
\description{
overload the addEpisode to patient.
}

