% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,ccRecord-method}
\alias{+,ccRecord,ccRecord-method}
\title{Combine two ccRecord objects}
\usage{
\S4method{+}{ccRecord,ccRecord}(e1, e2)
}
\description{
Combine two ccRecord objects
}

