% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,list-method}
\alias{+,ccRecord,list-method}
\title{Adding a list of ccEpisode to ccRecord}
\usage{
\S4method{+}{ccRecord,list}(e1, e2)
}
\arguments{
\item{e1}{ccRecord}

\item{e2}{a list of ccEpisode objects}
}
\value{
ccRecord
}
\description{
Adding a list of one or multiple ccEpisode objects to a
ccRecord object, the information table (infotb) will be updated automatically.
It is the more efficient way to add multiple ccEpisode objects.
}

