% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ccEpisode.r
\docType{methods}
\name{+,cEpisode,data.frame-method}
\alias{+,cEpisode,data.frame-method}
\title{add single or multiple record to episode.}
\usage{
\S4method{+}{cEpisode,data.frame}(e1, e2)
}
\description{
add single or multiple record to episode.
}

