% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sanity.check.r
\name{rate_range}
\alias{rate_range}
\title{when David's convert_type is merged, we will nolonger need to convert the
data type here.}
\usage{
rate_range(data, ref)
}
\arguments{
\item{data}{vector}

\item{ref}{result from yaml_range}
}
\description{
when David's convert_type is merged, we will nolonger need to convert the
data type here.
}

