% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readItems.r
\name{readOneItem}
\alias{readOneItem}
\title{Retrieve one item at a time from the ccRecord, NOTE: items with meta data,
i.e. three columns case is not valide for this function, should be improved
in the future.}
\usage{
readOneItem(record, item_id, convert_type = NULL, as.POSIX = FALSE,
  unlist = TRUE)
}
\arguments{
\item{record}{ccRecord}

\item{item_id}{NHIC code}
}
\description{
Retrieve one item at a time from the ccRecord, NOTE: items with meta data,
i.e. three columns case is not valide for this function, should be improved
in the future.
}

