% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{rec2tb}
\alias{rec2tb}
\title{Extract a variable from ccRecord object to a table.}
\usage{
rec2tb(ccd, varname)
}
\arguments{
\item{ccd}{ccRecord object}

\item{varname}{the short name of the selected variable, see ITEM_REF}
}
\description{
The output table will contain site_id, episode_id, time, val, and possible
meta data column if the variable is longitudinal. Otherwise only site_id,
episode_id and the value will be presented.
}

