% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{record2Table}
\alias{record2Table}
\title{Convert ccRecord data to item tables}
\usage{
record2Table(record, items = NULL, file = NULL)
}
\arguments{
\item{record}{ccRecord}

\item{items}{if specified, items should be the NHIC code of the selected
items, if not the output will contain be all the possible items, even
those not appearing in the ccRecord}

\item{file}{the name of csv files. If speicified 1d/2d tables will be
written into two sperate csv files.}
}
\description{
Convert ccRecord data to item tables
}

