% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccdata.r
\name{reindexRecord}
\alias{reindexRecord}
\title{Correct the index and meta data information for ccRecord.
It takes only the patient level indexing information.}
\usage{
reindexRecord(record)
}
\description{
Correct the index and meta data information for ccRecord.
It takes only the patient level indexing information.
}

