% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prefixID.r
\name{removeIdPrefix}
\alias{removeIdPrefix}
\title{remove the prefix of id, i.e. NIHR_HIC_ICU_0xxx -> xxx}
\usage{
removeIdPrefix(id)
}
\value{
character of numeric ids, e.g. "0123"
}
\description{
remove the prefix of id, i.e. NIHR_HIC_ICU_0xxx -> xxx
}

