% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repeatBinIndex.r
\name{repeatBinIndex}
\alias{repeatBinIndex}
\title{Provide an index of consequtively repeated items from a a given a binary vector.}
\usage{
repeatBinIndex(v)
}
\arguments{
\item{v}{vector only contains integer 0 or 1}
}
\value{
a vector without repeating.
}
\description{
Provide an index of consequtively repeated items from a a given a binary vector.
}

