% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectIndex.r
\name{selectIndex}
\alias{selectIndex}
\title{this one should no longer be useful.
here maybe some checking is necessary.}
\usage{
selectIndex(ids, type)
}
\value{
an selected index
}
\description{
this one should no longer be useful.
here maybe some checking is necessary.
}

