% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getItemList.r
\name{selectTable}
\alias{selectTable}
\title{getItemList}
\usage{
selectTable(record, items_opt = NULL, items_obg = NULL, freq,
  item.name = NULL)
}
\arguments{
\item{record}{ccRecord}

\item{items_opt}{optional items}

\item{items_obg}{obligatory items that is obligatory; Any episode that doesn't contain
item in this vector will be removed.}
}
\value{
data.table
}
\description{
getItemList
}

