% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_demographic.r
\name{sql.add.demographic}
\alias{sql.add.demographic}
\title{Add demographic table to the postgreSQL table.}
\usage{
sql.add.demographic(record, update = FALSE)
}
\description{
Add demographic table to the postgreSQL table.
}

