% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_demographic.R
\name{sql.demographic.table}
\alias{sql.demographic.table}
\title{Create demographic SQL tables. The data type of each column is in its
corresponding data type.}
\usage{
sql.demographic.table(record, dtype = TRUE)
}
\arguments{
\item{record}{ccRecord-class}

\item{dtype}{logical column will be type aware, else all in character.}
}
\description{
Create demographic SQL tables. The data type of each column is in its
corresponding data type.
}

