% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_demographic.r
\name{sql.init.postgres}
\alias{sql.init.postgres}
\title{Initialise the connection of between sqldf and postgres.}
\usage{
sql.init.postgres()
}
\description{
Initialise the connection of between sqldf and postgres.
}

