% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_demographic.r
\name{sql.load_file}
\alias{sql.load_file}
\title{Load SQL command from a file in which the SQL commands are recorded.}
\usage{
sql.load_file(path)
}
\arguments{
\item{path}{the full path of the SQL file.}
}
\value{
a string of command.
}
\description{
Load SQL command from a file in which the SQL commands are recorded.
}

