% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_demographic.r
\name{sql.newdb}
\alias{sql.newdb}
\title{Remove all the existing tables from the database and load the new schema
file.}
\usage{
sql.newdb(schema = NULL)
}
\arguments{
\item{schema}{is the full path of the selected schema file.}
}
\description{
Remove all the existing tables from the database and load the new schema
file.
}

