% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{StdId}
\alias{StdId}
\title{S4 class to hold standard IDs such as "NIHR_HIC_ICU_0001"}
\usage{
StdId(text)

StdId(text)
}
\arguments{
\item{text}{NIHC code which should be in a format like NIHR_HIC_ICU_xxxx}
}
\description{
S4 class to hold standard IDs such as "NIHR_HIC_ICU_0001"

constructor of StdId class
}
\section{Slots}{

\describe{
\item{\code{ids}}{single or multiple characters}
}}
\examples{
\donttest{
selected_tags <- StdId(c("NIHR_HIC_ICU_0001", "NIHR_HIC_ICU_0002")
}
}

