% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{stname2code}
\alias{stname2code}
\title{Convert short names to NHIC codes}
\usage{
stname2code(stname)
}
\arguments{
\item{stname}{character short names of data item h_rate}
}
\value{
NIHC code character such as NIHR_HIC_ICU_0108
}
\description{
Convert short names to NHIC codes
}

