% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{stname2longname}
\alias{stname2longname}
\title{Convert short names to long names.}
\usage{
stname2longname(stname)
}
\arguments{
\item{stname}{character short names of data item h_rate}
}
\value{
longname character such as "heart rate"
}
\description{
Convert short names to long names.
}

