% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{[,ccRecord,ANY-method}
\alias{[,ccRecord,ANY-method}
\title{Create a subset of ccRecord object from the original one via specifying the row number of episodes.}
\usage{
\S4method{[}{ccRecord,ANY}(x, i)
}
\arguments{
\item{x}{ccRecord-class}

\item{i}{integer vector}
}
\description{
Create a subset of ccRecord object from the original one via specifying the row number of episodes.
}

