% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{[,ccRecord,character-method}
\alias{[,ccRecord,character-method}
\title{Create a ccRecord subset via selected sites.}
\usage{
\S4method{[}{ccRecord,character}(x, i)
}
\description{
Create a ccRecord subset via selected sites.
}

