% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{subset,ccRecord,character-method}
\alias{subset,ccRecord,character-method}
\title{Subset episodes from the specified XML files.}
\usage{
\S4method{subset}{ccRecord,character}(r, f)
}
\arguments{
\item{ccd}{ccRecord object}

\item{files}{character a vector of XML file names - see ccRecord: parse_file}
}
\value{
ccRecord object
}
\description{
Subset episodes from the specified XML files.
}

