% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{subset}
\alias{subset}
\title{Subset episodes from the specified XML files.}
\usage{
subset(r, f)
}
\arguments{
\item{ccd}{ccRecord object}

\item{files}{character a vector of XML file names - see ccRecord: parse_file}
}
\value{
ccRecord object
}
\description{
Subset episodes from the specified XML files.
}

