% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{table1}
\alias{table1}
\title{Produce the item specified table one.}
\usage{
table1(demg, names, return.data = FALSE)
}
\arguments{
\item{demg}{ccTable-clas demographic table created by sql.demographic.table()}

\item{names}{character string. Short names of data items, e.g. h_rate.}

\item{return.data}{logical, FALSE: printing the pander table, TRUE: return the table but not print out the pander table.}
}
\value{
if return.data is TRUE, return data.table
}
\description{
Produce the item specified table one.
}

