% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{total.data.point}
\alias{total.data.point}
\title{Return total data point of the ccRecord object.}
\usage{
total.data.point(ccd)
}
\arguments{
\item{ccd}{ccRecord-class}
}
\description{
Return total data point of the ccRecord object.
}

