% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.spell.R
\name{unique_spell}
\alias{unique_spell}
\title{find the unique spell ID.}
\usage{
unique_spell(rec, duration = 2)
}
\arguments{
\item{rec}{ccRecord-class}

\item{duration}{integer hours}
}
\value{
data.table contains spell id.
}
\description{
find the unique spell ID.
}

