% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/unique.patient.r
\name{uniquePatients}
\alias{uniquePatients}
\title{Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)}
\usage{
uniquePatients(record)
}
\description{
Aggregate episodes with the same patient ids (e.g. nhs_number or pas_number)
}

