% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update.database}
\alias{update.database}
\title{updating the database}
\usage{
\method{update}{database}(xml.path, restart = FALSE, splitxml = FALSE,
  mc.cores = 4, quiet = FALSE)
}
\description{
something
}

