% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update.new.xml}
\alias{update.new.xml}
\title{update the RData database}
\usage{
\method{update}{new.xml}(xml.path, mc.cores = 4, quiet = FALSE)
}
\arguments{
\item{xml.path}{}

\item{mc.cores}{}
}
\value{
ccRecord object
combine new data to a ccRecord
}
\description{
Detect the new XML files which has never been parsed first and
inject the new episode data to the .RData database.
}

