% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update_database}
\alias{update_database}
\title{Update the critical care database (RData)}
\usage{
update_database(xml.path, restart = FALSE, splitxml = FALSE, mc.cores = 4,
  quiet = FALSE, dt = TRUE)
}
\arguments{
\item{xml.path}{character the path of the folder of which contains the XML files.}

\item{restart}{logical purge the previous database and restart parsing for all the XML files presented.}

\item{splitxml}{logical break down the XML files into chuncks. (Do it when the file is too big)}

\item{mc.cores}{integer number of processors to be applied for parallelisation.}

\item{quiet}{logical show the progress bar if true}
}
\description{
Parse critical care data from XML files and inject them into the RData
database.
}

