% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{which.classification}
\alias{which.classification}
\title{Identify the classification - classification1 of a given item code or short
name. Classification1 has 5 labels: 
[1] "Demographic", [2] "Physiology" 
[3] "Drugs" [4] "Nursing_other" [5] "Laboratory"}
\usage{
which.classification(item_name)
}
\arguments{
\item{item_name}{NHIC code or the short name}
}
\value{
character the item classification
}
\description{
Identify the classification - classification1 of a given item code or short
name. Classification1 has 5 labels: 
[1] "Demographic", [2] "Physiology" 
[3] "Drugs" [4] "Nursing_other" [5] "Laboratory"
}

