% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectIndex.r
\name{which.datatype}
\alias{which.datatype}
\title{Convert item data to its corresponding data type.}
\usage{
which.datatype(id, vals)
}
\arguments{
\item{id}{NHIC code of the data}

\item{vals}{The vector of values of the item.}
}
\value{
vector values in its corresponding data type
}
\description{
Convert item data to its corresponding data type.
}

