% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{whichIsCode}
\alias{whichIsCode}
\title{give id number from NHIC code like "NIHR_HIC_ICU_xxxx"}
\usage{
whichIsCode(nhic)
}
\arguments{
\item{nhic}{NHIC code}
}
\description{
give id number from NHIC code like "NIHR_HIC_ICU_xxxx"
}

