% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{xml.file.duration.plot}
\alias{xml.file.duration.plot}
\title{plot the duration of XML files.}
\usage{
xml.file.duration.plot(ccd)
}
\arguments{
\item{ccd}{ccRecord-class}
}
\description{
plot the duration of XML files.
}

