% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{xml.site.duration.plot}
\alias{xml.site.duration.plot}
\title{Plot the XML duration in terms of sites.}
\usage{
xml.site.duration.plot(ccd)
}
\arguments{
\item{ccd}{ccRecord-class}
}
\description{
Plot the XML duration in terms of sites.
}

