% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{xml2Data}
\alias{xml2Data}
\title{convert xml data to ccdata format}
\usage{
xml2Data(xml, select.episode = NULL, quiet = TRUE)
}
\arguments{
\item{xml}{can be either xml root or xml file name}
}
\value{
ccdata
}
\description{
convert xml data to ccdata format
}

