% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlEpisodeToList.r
\name{xmlEpisodeToList}
\alias{xmlEpisodeToList}
\title{Note: it rules out any time data with missing values.}
\usage{
xmlEpisodeToList(episode_node)
}
\description{
Note: it rules out any time data with missing values.
}

