% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cchic_xml.R
\name{xmlLoad}
\alias{xmlLoad}
\title{load xml clinical data}
\usage{
xmlLoad(file)
}
\arguments{
\item{file}{character string. The path of the XML file.}
}
\value{
the root of the xml data.
}
\description{
load xml clinical data
}
