% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlTime2POSIX.r
\name{xmlTime2POSIX}
\alias{xmlTime2POSIX}
\title{convert time from xml table to POSIX format.
e.g. 2014-02-01T03:00:00 -> 2014-02-01 03:00:00}
\usage{
xmlTime2POSIX(xml.time)
}
\description{
convert time from xml table to POSIX format.
e.g. 2014-02-01T03:00:00 -> 2014-02-01 03:00:00
}

