% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sanity.check.r
\name{yaml_range}
\alias{yaml_range}
\title{find the lower and upper bound of strings like "10 - 20"}
\usage{
yaml_range(ry)
}
\value{
c(lower_bound, upper_bound)
}
\description{
find the lower and upper bound of strings like "10 - 20"
}

