% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.episode.list.to.record}
\alias{add.episode.list.to.record}
\title{Adding a list of ccEpisode to ccRecord}
\usage{
add.episode.list.to.record(rec, lst)
}
\arguments{
\item{rec}{ccRecord}

\item{lst}{a list of ccEpisode objects}
}
\value{
ccRecord
}
\description{
Adding a list of one or multiple ccEpisode objects to a
ccRecord object, the information table (infotb) will be updated automatically.
It is the more efficient way to add multiple ccEpisode objects. See
add.episode.to.record() for just adding one ccEpisode.
}

