% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.episode.to.record}
\alias{add.episode.to.record}
\title{Adding one ccEpisode object to ccRecord object.}
\usage{
add.episode.to.record(rec, episode)
}
\arguments{
\item{rec}{ccRecord}

\item{episode}{ccEpisode object}
}
\value{
ccRecord object
}
\description{
Adding one ccEpisode object to ccRecord object.
}
\examples{
rec <- ccRecord()
eps <- new.episode()
new.rec <- add.episode.to.record(rec, eps)
# Alternatively
new.rec <- rec + eps
}

