% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.episode.to.record}
\alias{add.episode.to.record}
\title{Adding one ccEpisode object to ccRecord2 object.}
\usage{
add.episode.to.record(rec, episode)
}
\arguments{
\item{rec}{ccRecord2}

\item{episode}{ccEpisode2 object}
}
\value{
ccRecord2 object
}

