% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{add.record.to.record}
\alias{add.record.to.record}
\title{Concatenating two ccRecords objects}
\usage{
add.record.to.record(rec1, rec2)
}
\arguments{
\item{rec1}{ccRecord object}

\item{rec2}{ccRecord object}
}
\value{
ccRecord object
}

