% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{as.number}
\alias{as.number}
\title{convert standard ids to numbers (character) which can be used for indexing.}
\usage{
as.number(obj)
}
\description{
convert standard ids to numbers (character) which can be used for indexing.
}

