% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdid.R
\name{as.number}
\alias{as.number}
\title{Convert standard IDs to numbers (character) which can be used for indexing.}
\usage{
as.number(obj)
}
\arguments{
\item{obj}{a StdId object.}
}
\description{
Convert standard IDs to numbers (character) which can be used for indexing.
}

