% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\docType{class}
\name{ccTable-class}
\alias{ccTable}
\alias{ccTable-class}
\title{Rearrange 1d and 2d items into a wide table to enable data cleaning and
reporting process.}
\description{
Rearrange 1d and 2d items into a wide table to enable data cleaning and
reporting process.
}
\section{Fields}{

\describe{
\item{\code{record}}{ccRecord}

\item{\code{dfilter}}{list contains three data quality tables 1) range, 2)
missingness, 3) categorical}
}}
\section{Methods}{

\describe{
\item{\code{create.table(freq)}}{Create a table contains the selected items in the conf with a given
frequency (in hour)}

\item{\code{filter.category()}}{Check individual entries if they are the in the categories specified
in conf.}

\item{\code{filter.missingness(recount = FALSE)}}{filter out the where missingness is too low.}

\item{\code{filter.null(items = c("episode_id", "site"))}}{remove the entire episode when the episode_id or site is NULL}

\item{\code{reload.conf(file)}}{reload yaml configuration.}
}}

