% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaTime.R
\name{deltaTime}
\alias{deltaTime}
\title{convert calendar time data in a record to delta time comparing to the ICU
admission time.}
\usage{
deltaTime(record, pseudotime = FALSE, units = "hours", tdiff = FALSE)
}
\arguments{
\item{record}{ccRecord}

\item{pseudotime}{logical If pseudotime is set to be TRUE, then the
admission and discharge time will be set as the earliest and latest data stamp
in the record.}

\item{units}{units of delta time, which can be "hours", "mins", "days".}

\item{tdiff}{if false the delta time will be written in numeric format.}
}
\description{
convert calendar time data in a record to delta time comparing to the ICU
admission time.
}

