% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.table.R
\docType{methods}
\name{demographic.table}
\alias{demographic.table}
\title{demographic.table}
\usage{
demographic.table(obj)
}
\arguments{
\item{obj}{\code{\link{ccRecord-class}} or \code{\link{ccEpisode-class}}}
}
\value{
return a data.table with all demographic fields of each episode.
}
\description{
demographic.table
}
\keyword{methods}

