% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.r
\name{duplicated.episode}
\alias{duplicated.episode}
\title{pick up the identical episode which has been injected by mistake.}
\usage{
\method{duplicated}{episode}(cr)
}
\description{
pick up the identical episode which has been injected by mistake.
}

