% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocateTime.R
\name{getEpisodePeriod}
\alias{getEpisodePeriod}
\title{getEpisodePeriod}
\usage{
getEpisodePeriod(e, unit = "hours")
}
\arguments{
\item{e}{episode}
}
\value{
period_length
}
\description{
getEpisodePeriod
}

