% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reallocateTime.R
\name{getEpisodePeriod}
\alias{getEpisodePeriod}
\title{Get the length of stay based on the first and the last data point.}
\usage{
getEpisodePeriod(e, unit = "hours")
}
\arguments{
\item{e}{ccEpisode object.}

\item{unit}{character string.  Units in which the results are desired. Can be abbreviated.}
}
\value{
length of stay
}
\description{
Get the length of stay based on the first and the last data point.
}

