% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccTable.R
\name{getfilter}
\alias{getfilter}
\title{get the dfilter}
\usage{
getfilter(dq, criterion)
}
\arguments{
\item{dq}{can be either dqaulity table or torigin}

\item{criterion}{should be a function to give T/F values of each entry.}
}
\description{
get the dfilter
}

