% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{getXmlepisode}
\alias{getXmlepisode}
\title{get the episode data from xml}
\usage{
getXmlepisode(xml.root, id)
}
\arguments{
\item{xml.root}{root of xml data returned by xmlLoad()}
}
\value{
episode xml data
}
\description{
get the episode data from xml
}

