% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2ccdata.R
\name{getXmlepisode}
\alias{getXmlepisode}
\title{get the episode data from xml}
\usage{
getXmlepisode(xml.root, id)
}
\arguments{
\item{xml.root}{root of xml data returned by xmlLoad()}

\item{id}{integer}
}
\description{
get the episode data from xml
}

