% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.range.R
\name{inrange}
\alias{inrange}
\title{Check if the values of a vector v is in the given ranges.}
\usage{
inrange(v, range)
}
\arguments{
\item{v}{vector numeric}

\item{range}{A string contains the numeric ranges in a form such as (low,
up) for open range and [low, up] for close range. Multiple
ranges should be separated by semi-columns which is equivalent to logical
OR. e.g. (low1, up1); (low2, up2)}
}
\description{
Check if the values of a vector v is in the given ranges.
}

