% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.r
\name{interpolateVec}
\alias{interpolateVec}
\title{perform window period approximation on a given numeric vector. Window is
specified by lead and lag.}
\usage{
interpolateVec(v, lead, lag, FUN = "mean", ...)
}
\arguments{
\item{v}{vector}

\item{lead}{number of forward element from the centre of the window.}
}
\description{
perform window period approximation on a given numeric vector. Window is
specified by lead and lag.
}

