% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{interpolateVec}
\alias{interpolateVec}
\title{Numeric vector interpolation}
\usage{
interpolateVec(v, lead, lag, FUN = "mean", ...)
}
\arguments{
\item{v}{vector}

\item{lead}{number of forward element from the missing value.}

\item{lag}{number of element backward from the missing value.}

\item{FUN}{the interpolation function.}
}
\value{
vector
}
\details{
Performing interpolation methods on a given window which is
specified by lead and lag.
}

