% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R, R/utilities.R
\name{lookup.items}
\alias{lookup.items}
\alias{lookup.items}
\title{Lookup codes based by keywords}
\usage{
lookup.items(keyword, style = "grid")

lookup.items(keyword, style = "grid")
}
\arguments{
\item{keyword}{character e.g. "heart", "108". 
If you provide multiple keywords it will try match all of them.}

\item{style}{character, the style of the table output which can be "simple",
"rmarkdown", and "grid"}

\item{metadata}{The metadata table. This must be the actual object (collect()ed, not a db connection)}

\item{keyword}{character e.g. "h_rate", "heart", "108".}
}
\value{
A subset of the metadatatable which contains only rows which have the given keyword.

character the short names of the selected items.
}
\description{
This function tries to match keywords.
The matched items will be displayed.

This function tries to match keywords in short names, long names and NHIC code. 
The matched items will be displayed.
}
\examples{
lookup.items('heart')
lookup.items(c('admission', 'unit'))
}
