% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbConnection.R
\name{metadata}
\alias{metadata}
\title{The metadata table. This is all of the different 
codes that are available with their meanings, and
what the different columns mean for them.}
\usage{
metadata(connection)
}
\description{
The metadata table explains the meanings of data in
the events table, which is retrieved through the
exportData() function. It has three fields to describe
what the general shape of the data is, while the other
fields describe what data is stored in the field with that 
name in the events table.
}
\details{
* code_name is the XML code of this type of data
* long_name is an english description of what it is
* primary_column is what column of the events table 
     stores the observation value for this data type
* All others: a plain english description of what is stored 
     in that column in the events table. NA if nothing.

This function returns a loaded into memory R table,
**not** a connection to database table.
}

