% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{physio.distribution}
\alias{physio.distribution}
\title{Plot the physiological data distribution.}
\usage{
physio.distribution(cctb, names)
}
\arguments{
\item{cctb}{ccTable-class, see create.cctable().}

\item{names}{character vector of short names of numerical demographic data.}
}
\description{
Plot the physiological data distribution.
}

