% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,ccEpisode-method}
\alias{+,ccRecord,ccEpisode-method}
\title{Adding one ccEpisode object to a ccRecord}
\usage{
\S4method{+}{ccRecord,ccEpisode}(e1, e2)
}
\description{
Adding one ccEpisode object to a ccRecord
}

