% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,ccEpisode-method}
\alias{+,ccRecord,ccEpisode-method}
\title{Adding one ccEpisode object to a ccRecord}
\usage{
\S4method{+}{ccRecord,ccEpisode}(e1, e2)
}
\arguments{
\item{e1}{ccRecord-class}

\item{e2}{ccEpisode-class}
}
\value{
ccRecord-class
}
\description{
Adding one ccEpisode object to a ccRecord
}

