% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,list-method}
\alias{+,ccRecord,list-method}
\title{Adding a list of ccEpisode objects to a ccRecord}
\usage{
\S4method{+}{ccRecord,list}(e1, e2)
}
\description{
Adding a list of ccEpisode objects to a ccRecord
}

