% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,list-method}
\alias{+,ccRecord,list-method}
\title{Adding a list of ccEpisode objects to a ccRecord}
\usage{
\S4method{+}{ccRecord,list}(e1, e2)
}
\arguments{
\item{e1}{ccRecord-class}

\item{e2}{A list of ccEpisode objects}
}
\value{
ccRecord-class
}
\description{
Adding a list of ccEpisode objects to a ccRecord
}

