% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{+,ccRecord,NULL-method}
\alias{+,ccRecord,NULL-method}
\title{Adding nothing to a ccRecord object.}
\usage{
\S4method{+}{ccRecord,`NULL`}(e1, e2)
}
\arguments{
\item{e1}{ccRecord-class}

\item{e2}{NULL}
}
\description{
Adding nothing to a ccRecord object.
}

