% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_data.r
\name{readOneItem}
\alias{readOneItem}
\title{Retrieve one item at a time from the ccRecord}
\usage{
readOneItem(record, item_id, convert_type = NULL, as.POSIX = FALSE,
  unlist = TRUE)
}
\arguments{
\item{record}{ccRecord}

\item{item_id}{NHIC code}
}
\description{
Retrieve one item at a time from the ccRecord
}

