% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{recreate.env}
\alias{recreate.env}
\title{This is a patch for reloading environment variables after doing clean all.}
\usage{
recreate.env()
}
\description{
This is a patch for reloading environment variables after doing clean all.
}

