% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.quality.report.R
\name{samplerate2d}
\alias{samplerate2d}
\title{Produce a pander table of sample rate of longitudinal data.}
\usage{
samplerate2d(cctb)
}
\arguments{
\item{cctb}{ccTable-class, see create.cctable().}
}
\description{
Produce a pander table of sample rate of longitudinal data.
}

