% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{select_data}
\alias{select_data}
\title{select one data item from ccRecord by unit by time}
\usage{
select_data(record, item, sites = NULL, years = NULL, propgt = FALSE,
  freq = NULL)
}
\description{
select one data item from ccRecord by unit by time
}

