% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_data.r
\name{select_data}
\alias{select_data}
\title{select one data item from ccRecord by unit by time}
\usage{
select_data(record, item, sites = NULL, years = NULL, propagate = FALSE,
  frequency = NULL)
}
\description{
select one data item from ccRecord by unit by time
}

