% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectIndex.r
\name{selectIndex}
\alias{selectIndex}
\title{here maybe some checking is necessary.}
\usage{
selectIndex(info.index, ids, type)
}
\value{
an selected index
}
\description{
here maybe some checking is necessary.
}

