% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTable.R
\name{selectTable}
\alias{selectTable}
\title{selectTable}
\usage{
selectTable(record, items_opt = NULL, items_obg = NULL, freq,
  item.name = NULL, return_list = FALSE)
}
\arguments{
\item{record}{ccRecord}

\item{items_opt}{optional items}

\item{items_obg}{obligatory items that is obligatory; Any episode that doesn't contain
item in this vector will be removed.}
}
\value{
data.table
}
\description{
selectTable
}

