% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{[,ccRecord,character-method}
\alias{[,ccRecord,character-method}
\title{Create a ccRecord subset via selected sites.}
\usage{
\S4method{[}{ccRecord,character}(x, i)
}
\arguments{
\item{x}{ccRecord-class}

\item{i}{character vector which contains site_ids, e.g. c("Q70", "Q70W")}
}
\description{
Create a ccRecord subset via selected sites.
}

