% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{[[,ccRecord-method}
\alias{[[,ccRecord-method}
\title{Subsetting a ccRecord object and return a list of ccEpisode objects.}
\usage{
\S4method{[[}{ccRecord}(x, i)
}
\arguments{
\item{x}{ccRecord-class}

\item{i}{integer vector}
}
\description{
Subsetting a ccRecord object and return a list of ccEpisode objects.
}
