% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\docType{methods}
\name{subset,ccRecord,character-method}
\alias{subset,ccRecord,character-method}
\title{Get a subset of episodes that have the same from ccRecord .}
\usage{
\S4method{subset}{ccRecord,character}(r, f)
}
\arguments{
\item{r}{ccRecord-class}

\item{f}{character a vector of XML file names - see ccRecord: parse_file}
}
\value{
ccRecord object
}
\description{
Get a subset of episodes that have the same from ccRecord .
}

