% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccRecord.R
\name{subset}
\alias{subset}
\title{Get a subset of episodes from ccRecord.}
\usage{
subset(r, f)
}
\arguments{
\item{r}{ccRecord-class}

\item{f}{character a vector of XML file names - see ccRecord: parse_file}
}
\value{
ccRecord-class
}
\description{
Get a subset of episodes from ccRecord.
}

