% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.spell.R
\name{unique.spell}
\alias{unique.spell}
\title{find the unique spell ID.}
\usage{
\method{unique}{spell}(rec, duration = 2)
}
\arguments{
\item{record}{ccRecord}
}
\value{
data.table contains spell id.
}
\description{
find the unique spell ID.
}

