% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{update.new.xml}
\alias{update.new.xml}
\title{Update the RData database}
\usage{
\method{update}{new.xml}(xml.path, mc.cores = 4, quiet = FALSE)
}
\arguments{
\item{mc.cores}{}
}
\value{
ccRecord object
combine new data to a ccRecord
}
\description{
Inject episode data from the newly added XML files to the RData database.
}

