% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{xml2Data}
\alias{xml2Data}
\title{convert xml data to ccdata format}
\usage{
xml2Data(file, select.episode = NULL, quiet = TRUE, xml = NULL,
  file_origin = "NA", parse_time = Sys.time())
}
\arguments{
\item{file}{xml file name}
}
\value{
ccdata
}
\description{
convert xml data to ccdata format
}

