% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xml2ccdata.r
\name{xml2Data}
\alias{xml2Data}
\title{convert xml data to ccdata format}
\usage{
xml2Data(xml, select.patient = NULL)
}
\arguments{
\item{xml}{xml root}
}
\value{
ccdata
}
\description{
convert xml data to ccdata format
}

