% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2ccdata.R
\name{xmlEpisodeToList}
\alias{xmlEpisodeToList}
\title{Note: it rules out any time data with missing values.}
\usage{
xmlEpisodeToList(episode_node)
}
\description{
Note: it rules out any time data with missing values.
}

