% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{xmlTime2POSIX}
\alias{xmlTime2POSIX}
\title{Convert time from xml to POSIX format.}
\usage{
xmlTime2POSIX(xml.time, allow = FALSE)
}
\arguments{
\item{xml.time}{character. Time in XML format such as 2014-02-01T03:00:00}

\item{allow}{logical. Wrong format will be accepted when \code{allow} is set
to be TRUE and NA will be the return value, otherwise return error. 
It is useful while dealing with pseudonymous data where the time format is
not presented correctly.}
}
\description{
Convert the XML time The XML time format to POSIXct.
}

